fix(extension): [lw-10539] add isSharedWallet key on wallet slive level
