+++ apps/browser-extension-wallet/src/stores/slices/wallet-info-slice.ts
@@ -2,6 +2,7 @@ import { WalletType } from '@cardano-sdk/web-extension';
 import { BlockchainProviderSlice, SliceCreator, WalletInfoSlice } from '../types';
 import { Wallet } from '@lace/cardano';
 import { ObservableWalletState } from '@hooks/useWalletState';
+import { isSharedWallet } from '@src/utils/is-shared-wallet';
 
 /**
  * has all wallet info related actions and states
@@ -23,6 +24,7 @@ export const walletInfoSlice: SliceCreator<WalletInfoSlice & BlockchainProviderS
   initialHdDiscoveryCompleted: false,
   isInMemoryWallet: undefined,
   isHardwareWallet: undefined,
+  isSharedWallet: undefined,
   walletType: undefined,
   stayOnAllDonePage: false,
   setAddressesDiscoveryCompleted: (addressesDiscoveryCompleted) =>
@@ -36,7 +38,8 @@ export const walletInfoSlice: SliceCreator<WalletInfoSlice & BlockchainProviderS
       cardanoWallet: wallet,
       walletType: wallet?.source.wallet.type,
       isInMemoryWallet: wallet?.source.wallet.type === WalletType.InMemory,
-      isHardwareWallet: [WalletType.Ledger, WalletType.Trezor].includes(wallet?.source.wallet.type)
+      isHardwareWallet: [WalletType.Ledger, WalletType.Trezor].includes(wallet?.source.wallet.type),
+      isSharedWallet: isSharedWallet(wallet)
     }),
   setCurrentChain: (chain: Wallet.ChainName) => {
     set({ currentChain: Wallet.Cardano.ChainIds[chain], environmentName: chain });
